improve: change folder structure, add improvements to product page
