+++ src/app/catalog/[category]/consts.ts
@@ -39,13 +39,5 @@ export const CATEGORIES: Category[] = [
   { text: "Верхняя одежда", href: "/" },
   { text: "Бестселлеры", href: "/" },
   { text: "Коллекции", href: "/" },
-  { text: "Биба", href: "/" },
-  { text: "Буба", href: "/" },
-  { text: "быба", href: "/" },
-  { text: "Боба", href: "/" },
-  { text: "бева", href: "/" },
-  { text: "ьыва", href: "/" },
-  { text: "тыва", href: "/" },
-  { text: "ыва", href: "/" },
-  { text: "кыва", href: "/" },
+  { text: "Final Sale", href: "/" },
 ];
