+++ src/app/catalog/[category]/page.tsx
