+++ src/app/catalog/product/components/AboutSection/AboutSection.tsx
