+++ src/app/catalog/product/components/ProductSection/ProductSection.tsx
