+++ src/app/catalog/product/components/ProductSection/ProductSlider/ProductSlider.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import { MouseEvent, useState } from "react";
+import { useState } from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 
 import Image from "next/image";
-import ImageWithFallback from "@/components/UI/ImageWithFallback";
+import ImageWithZoom from "@/components/UI/ImageWithZoom";
 
 import { Thumbs } from "swiper/modules";
 import { Swiper as SwiperType } from "swiper/types";
@@ -27,19 +27,6 @@ const SLIDES = [
 const ProductSlider = () => {
   const [thumbsSwiper, setThumbsSwiper] = useState<SwiperType | null>(null);
 
-  const [zoom, setZoom] = useState({ x: "50%", y: "50%", scale: 1 });
-
-  const handleMouseMove = (e: MouseEvent<HTMLDivElement>) => {
-    const { left, top, width, height } =
-      e.currentTarget.getBoundingClientRect();
-    const x = ((e.clientX - left) / width) * 100;
-    const y = ((e.clientY - top) / height) * 100;
-
-    setZoom({ x: `${x}%`, y: `${y}%`, scale: 1.5 });
-  };
-
-  const resetZoom = () => setZoom((prev) => ({ ...prev, scale: 1 }));
-
   return (
     <div className="product-slider flex gap-6 flex-1 basis-[51%] overflow-hidden">
       <Swiper
@@ -59,7 +46,7 @@ const ProductSlider = () => {
             <Image
               fill
               src={slide}
-              className="object-cover"
+              className="object-cover cursor-pointer"
               alt="product-slide-thumb"
             />
           </SwiperSlide>
@@ -68,28 +55,13 @@ const ProductSlider = () => {
 
       <Swiper
         className="main-slider"
+        allowTouchMove={false}
         thumbs={{ swiper: thumbsSwiper }}
         modules={[Thumbs]}
       >
         {SLIDES.map((slide) => (
-          // Ключ нужно будет сделать уникальным
           <SwiperSlide key={slide}>
-            <div
-              className="zoom-container"
-              onMouseMove={handleMouseMove}
-              onMouseLeave={resetZoom}
-            >
-              <ImageWithFallback
-                src={slide}
-                className=" transition-transform w-full h-auto"
-                fallbackSrc={slide}
-                style={{
-                  transformOrigin: `${zoom.x} ${zoom.y}`,
-                  transform: `scale(${zoom.scale})`,
-                }}
-                alt="product-image"
-              />
-            </div>
+            <ImageWithZoom src={slide} />
           </SwiperSlide>
         ))}
       </Swiper>
