+++ src/app/catalog/product/components/ProductSection/ProductSlider/styles.scss
