+++ src/components/UI/Progress.tsx
@@ -12,7 +12,7 @@ const Progress = React.forwardRef<
   <ProgressPrimitive.Root
     ref={ref}
     className={cn(
-      "relative h-1 w-full overflow-hidden rounded-full bg-disabled",
+      "relative h-1 w-full overflow-hidden rounded-full bg-gray-light",
       className
     )}
     {...props}
